Move service module state to a separate file
